pyproject.toml
[build-system]
requires = ["hatchling>=1.22.0"]
build-backend = "hatchling.build"

[project]
name = "llmwalk"
version = "0.1.3"
description = "Explore the answer-space for any prompt and any MLX-supported model."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "mlx-lm==0.29.1",
  "rich==14.2.0",
  "sortedcontainers==2.4.0",
  "transformers===4.57.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "syrupy>=4.0.0",
]

[project.scripts]
llmwalk = "llmwalk.cli:main"

[tool.hatch.build]
exclude = [
  "example1.gif",
]

[tool.hatch.build.targets.wheel]
packages = ["llmwalk"]

[tool.pytest.ini_options]
filterwarnings = [
  # Noise from the sentencepiece SWIG extension when running with Mistral tokenizers.
  "ignore:.*SwigPyPacked.*:DeprecationWarning",
  "ignore:.*SwigPyObject.*:DeprecationWarning",
  "ignore:.*swigvarlink.*:DeprecationWarning",
]
